## Description

- in 'main' splits off cruise and format in their own modules
- in the new 'cruise' module parallelizes dynamic imports of everything that can be optionally imported
- in the new 'cruise' module dynamically load _everything_ that can be optionally imported (in stead of just the perceived 'biggest wins')

## Motivation and Context

- the split improves legibility/ maintainability
- the rest improves performance (normal runs without cache it doesn't make much difference;  cached, and when it can served directly from cache ~60ms)

## How Has This Been Tested?

- [x] green ci

## Comparisons

### Before
#### serving from cache
```
$ hyperfine --warmup 3 --runs 30 "bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --cache node_modules/.cache/dependency-cruiser/load-cached --no-progress"
Benchmark 1: bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --cache node_modules/.cache/dependency-cruiser/load-cached --no-progress
  Time (mean ± σ):      1.262 s ±  0.012 s    [User: 1.320 s, System: 0.197 s]
  Range (min … max):    1.248 s …  1.299 s    30 runs
```
#### regular run

```
Benchmark 1: bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --no-cache --no-progress
  Time (mean ± σ):      3.068 s ±  0.019 s    [User: 4.715 s, System: 0.318 s]
  Range (min … max):    3.041 s …  3.118 s    30 runs
```
### After

#### serving from cache
```
$ hyperfine --warmup 3 --runs 30 "bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --cache node_modules/.cache/dependency-cruiser/load-cached --no-progress"
Benchmark 1: bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --cache node_modules/.cache/dependency-cruiser/load-cached --no-progress
  Time (mean ± σ):      1.200 s ±  0.013 s    [User: 1.229 s, System: 0.189 s]
  Range (min … max):    1.184 s …  1.236 s    30 runs
```
#### regular run
```
$ hyperfine --warmup 3 --runs 30 "bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --no-cache --no-progress"
Benchmark 1: bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --no-cache --no-progress
  Time (mean ± σ):      3.066 s ±  0.019 s    [User: 4.761 s, System: 0.311 s]
  Range (min … max):    3.043 s …  3.130 s    30 runs
```

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation only change
- [ ] Refactor (non-breaking change which fixes an issue without changing functionality)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

- [ ] :book:

  - My change doesn't require a documentation update, or ...
  - it _does_ and I have updated it

- [ ] :balance_scale:
  - The contribution will be subject to [The MIT license](https://github.com/sverweij/dependency-cruiser/blob/develop/LICENSE), and I'm OK with that.
  - The contribution is my own original work.
  - I am ok with the stuff in [**CONTRIBUTING.md**](https://github.com/sverweij/dependency-cruiser/blob/develop/.github/CONTRIBUTING.md).